+++ src/frontend/app/shared-tdoc/components/tdoc-linked-infos/tdoc-linked-infos.component.spec.ts
@@ -1,0 +1,47 @@
+/* tslint:disable:no-unused-variable */
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+import {NO_ERRORS_SCHEMA} from '@angular/core';
+import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
+import {TranslateModule} from '@ngx-translate/core';
+import {TourDocDataServiceStub} from '../../../../testing/tdoc-dataservice-stubs';
+import {Router} from '@angular/router';
+import {RouterStub} from '@dps/mycms-frontend-commons/dist/angular-commons/testing/router-stubs';
+import {CommonRoutingService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/common-routing.service';
+import {CommonDocRoutingService} from '@dps/mycms-frontend-commons/dist/frontend-cdoc-commons/services/cdoc-routing.service';
+import {TourDocContentUtils} from '../../services/tdoc-contentutils.service';
+import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
+import {AppServiceStub} from '@dps/mycms-frontend-commons/dist/angular-commons/testing/appservice-stubs';
+import {TourDocLinkedInfosComponent} from './tdoc-linked-infos.component';
+
+describe('TourDocLinkedInfosComponent', () => {
+    let component: TourDocLinkedInfosComponent;
+    let fixture: ComponentFixture<TourDocLinkedInfosComponent>;
+
+    beforeEach(async(() => {
+        TestBed.configureTestingModule({
+            declarations: [TourDocLinkedInfosComponent],
+            providers: [
+                { provide: Router, useValue: new RouterStub() },
+                CommonRoutingService,
+                CommonDocRoutingService,
+                TourDocContentUtils,
+                { provide: GenericAppService, useValue: new AppServiceStub() }
+            ],
+            schemas: [NO_ERRORS_SCHEMA],
+            imports: [NgbModule,
+                TranslateModule.forRoot()]
+        })
+            .compileComponents();
+    }));
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(TourDocLinkedInfosComponent);
+        component = fixture.componentInstance;
+        component.record = TourDocDataServiceStub.defaultRecord();
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+});